+++ changelog.md
@@ -5,6 +5,7 @@
 * Nerfed Unbreaking (items last ~~50%/120%/190%/260%~~ -> ~~40%/110%/180%/250%~~ more)
 * Hopefully fixed Farmers' Delight wild carrots, potatoes and beetroots from generating
 * Can no longer smelt Farmers' Delight stuff in furnace with no_food_in_furnace data pack enabled
+* Reduced Sweet Berry Meatballs and Glowgurt nutrition
 
 ## Beta 3.9.2
 * Broken items now show a red durability bar